SingleTheme class

SingleTheme class

SingleTheme class

Details

Definition for the SingleTheme class.

The relative order is fixed at 1 because each this class only contains a single feature.

See also

Super class

wheretowork::Theme -> SingleTheme

Methods

Public methods

SingleTheme$new()

SingleTheme$get_feature_order()

SingleTheme$set_feature_order()

SingleTheme$get_solution_settings_widget_data()

SingleTheme$get_map_manager_widget_data()

SingleTheme$clone()

Inherited methods

wheretowork::Theme$export()

wheretowork::Theme$get_feature_current()

wheretowork::Theme$get_feature_goal()

wheretowork::Theme$get_feature_id()

wheretowork::Theme$get_feature_limit()

wheretowork::Theme$get_feature_name()

wheretowork::Theme$get_feature_status()

wheretowork::Theme$get_feature_total()

wheretowork::Theme$get_feature_visible()

wheretowork::Theme$get_layer_index()

wheretowork::Theme$get_layer_name()

wheretowork::Theme$get_setting()

wheretowork::Theme$get_visible()

wheretowork::Theme$print()

wheretowork::Theme$render_on_map()

wheretowork::Theme$repr()

wheretowork::Theme$set_feature_current()

wheretowork::Theme$set_feature_goal()

wheretowork::Theme$set_feature_status()

wheretowork::Theme$set_feature_visible()

wheretowork::Theme$set_setting()

wheretowork::Theme$set_visible()

wheretowork::Theme$update_on_map()

Method new()

Create a SingleTheme object.

Usage

SingleTheme$new(id, name, feature)

Arguments

id

character value.

name

character value.

feature

list of a single Feature object.

Returns

A new SingleTheme object.

Method new()

Create a SingleTheme object.

Usage

SingleTheme$new(id, name, feature)

Arguments

id

character value.

name

character value.

feature

list of a single Feature object.

Returns

A new SingleTheme object.

Method get_feature_order()

Get relative order for displaying each feature on a map.

Usage

SingleTheme$get_feature_order()

Returns

numeric value.

Method get_feature_order()

Get relative order for displaying each feature on a map.

Usage

SingleTheme$get_feature_order()

Returns

numeric value.

Method set_feature_order()

Set relative order for displaying features on a map.

Usage

SingleTheme$set_feature_order(...)

Arguments

...

not used.

Method set_feature_order()

Set relative order for displaying features on a map.

Usage

SingleTheme$set_feature_order(...)

Arguments

...

not used.

Method get_solution_settings_widget_data()

Get data for displaying the theme in a solutionSettings widget.

Usage

SingleTheme$get_solution_settings_widget_data()

Returns

list with widget data.

Method get_solution_settings_widget_data()

Get data for displaying the theme in a solutionSettings widget.

Usage

SingleTheme$get_solution_settings_widget_data()

Returns

list with widget data.

Method get_map_manager_widget_data()

Get data for displaying the theme in a mapManager() widget.

Usage

SingleTheme$get_map_manager_widget_data()

Returns

list with widget data.

Method get_map_manager_widget_data()

Get data for displaying the theme in a mapManager() widget.

Usage

SingleTheme$get_map_manager_widget_data()

Returns

list with widget data.

Method clone()

The objects of this class are cloneable with this method.

Usage

SingleTheme$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Method clone()

The objects of this class are cloneable with this method.

Usage

SingleTheme$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.