Theme class

Theme class

Theme class

Details

Definition for the Theme class. This is a parent class that contains shared fields and methods for the SingleTheme and MultiTheme classes.

This method intentionally throws an error.

See also

Public fields

id

character value.

name

character value.

feature

list of Feature objects.

Methods

Public methods

Theme$new()

Theme$print()

Theme$repr()

Theme$get_layer_name()

Theme$get_layer_index()

Theme$get_feature_id()

Theme$get_feature_name()

Theme$get_feature_current()

Theme$get_feature_limit()

Theme$get_feature_total()

Theme$get_visible()

Theme$get_feature_visible()

Theme$get_feature_status()

Theme$get_feature_goal()

Theme$get_setting()

Theme$set_visible()

Theme$set_feature_visible()

Theme$set_feature_status()

Theme$set_feature_goal()

Theme$set_feature_current()

Theme$set_setting()

Theme$export()

Theme$render_on_map()

Theme$update_on_map()

Theme$clone()

Method new()

Create a Theme object.

Usage

Theme$new(...)

Arguments

...

not used.

Method new()

Create a Theme object.

Usage

Theme$new(...)

Arguments

...

not used.

Method print()

Print the object.

Usage

Theme$print(...)

Arguments

...

not used.

Method print()

Print the object.

Usage

Theme$print(...)

Arguments

...

not used.

Method repr()

Generate a character summarizing the representation of the object.

Usage

Theme$repr(...)

Arguments

...

not used.

Returns

character value.

Method repr()

Generate a character summarizing the representation of the object.

Usage

Theme$repr(...)

Arguments

...

not used.

Returns

character value.

Method get_layer_name()

Get layer names.

Usage

Theme$get_layer_name()

Returns

character vector.

Method get_layer_name()

Get layer names.

Usage

Theme$get_layer_name()

Returns

character vector.

Method get_layer_index()

Get layer index values.

Usage

Theme$get_layer_index()

Returns

character vector.

Method get_layer_index()

Get layer index values.

Usage

Theme$get_layer_index()

Returns

character vector.

Method get_feature_id()

Get feature identifiers.

Usage

Theme$get_feature_id()

Returns

character vector with identifier(s).

Method get_feature_id()

Get feature identifiers.

Usage

Theme$get_feature_id()

Returns

character vector with identifier(s).

Method get_feature_name()

Get feature names.

Usage

Theme$get_feature_name()

Returns

character vector with identifier(s).

Method get_feature_name()

Get feature names.

Usage

Theme$get_feature_name()

Returns

character vector with identifier(s).

Method get_feature_current()

Get feature current.

Usage

Theme$get_feature_current()

Returns

numeric vector with value(s).

Method get_feature_current()

Get feature current.

Usage

Theme$get_feature_current()

Returns

numeric vector with value(s).

Method get_feature_limit()

Get feature limit.

Usage

Theme$get_feature_limit()

Returns

numeric vector with value(s).

Method get_feature_limit()

Get feature limit.

Usage

Theme$get_feature_limit()

Returns

numeric vector with value(s).

Method get_feature_total()

Get feature current.

Usage

Theme$get_feature_total()

Returns

numeric vector with value(s).

Method get_feature_total()

Get feature current.

Usage

Theme$get_feature_total()

Returns

numeric vector with value(s).

Method get_visible()

Set visible value for all features.

Usage

Theme$get_visible()

Returns

logical value.

Method get_visible()

Set visible value for all features.

Usage

Theme$get_visible()

Returns

logical value.

Method get_feature_visible()

Get feature visible values.

Usage

Theme$get_feature_visible()

Returns

logical vector with status value(s).

Method get_feature_visible()

Get feature visible values.

Usage

Theme$get_feature_visible()

Returns

logical vector with status value(s).

Method get_feature_status()

Get feature status values.

Usage

Theme$get_feature_status()

Returns

logical vector with status value(s).

Method get_feature_status()

Get feature status values.

Usage

Theme$get_feature_status()

Returns

logical vector with status value(s).

Method get_feature_goal()

Get feature goal values.

Usage

Theme$get_feature_goal()

Returns

numeric vector with goal value(s).

Method get_feature_goal()

Get feature goal values.

Usage

Theme$get_feature_goal()

Returns

numeric vector with goal value(s).

Method get_setting()

Get setting.

Usage

Theme$get_setting(name)

Arguments

name

character setting name. Available options are "feature_status", "feature_goal", "feature_visible", "visible", or "feature_order".

Returns

Value.

Method get_setting()

Get setting.

Usage

Theme$get_setting(name)

Arguments

name

character setting name. Available options are "feature_status", "feature_goal", "feature_visible", "visible", or "feature_order".

Returns

Value.

Method set_visible()

Set visible value for all features.

Usage

Theme$set_visible(value)

Arguments

value

logical value.

Method set_visible()

Set visible value for all features.

Usage

Theme$set_visible(value)

Arguments

value

logical value.

Method set_feature_visible()

Set feature visible values.

Usage

Theme$set_feature_visible(value)

Arguments

value

logical vector containing a value for each feature. A list of logical values can also be supplied.

Method set_feature_visible()

Set feature visible values.

Usage

Theme$set_feature_visible(value)

Arguments

value

logical vector containing a value for each feature. A list of logical values can also be supplied.

Method set_feature_status()

Set feature status values.

Usage

Theme$set_feature_status(value)

Arguments

value

logical vector containing a value for each feature. A list of logical values can also be supplied.

Method set_feature_status()

Set feature status values.

Usage

Theme$set_feature_status(value)

Arguments

value

logical vector containing a value for each feature. A list of logical values can also be supplied.

Method set_feature_goal()

Set feature goal values.

Usage

Theme$set_feature_goal(value)

Arguments

value

numeric vector containing a value for each feature. A list of numeric values can also be supplied.

Method set_feature_goal()

Set feature goal values.

Usage

Theme$set_feature_goal(value)

Arguments

value

numeric vector containing a value for each feature. A list of numeric values can also be supplied.

Method set_feature_current()

Set feature current values.

Usage

Theme$set_feature_current(value)

Arguments

value

numeric vector containing a value for each feature. A list of numeric values can also be supplied.

Method set_feature_current()

Set feature current values.

Usage

Theme$set_feature_current(value)

Arguments

value

numeric vector containing a value for each feature. A list of numeric values can also be supplied.

Method set_setting()

Set setting.

Usage

Theme$set_setting(name, value)

Arguments

name

character setting name. Available options are "feature_status", "feature_goal", "feature_visible", "visible", "feature_order", or "feature_current".

value

vector containing a value for each feature.

Method set_setting()

Set setting.

Usage

Theme$set_setting(name, value)

Arguments

name

character setting name. Available options are "feature_status", "feature_goal", "feature_visible", "visible", "feature_order", or "feature_current".

value

vector containing a value for each feature.

Method export()

Export settings

Usage

Theme$export()

Returns

list object.

Method export()

Export settings

Usage

Theme$export()

Returns

list object.

Method render_on_map()

Render on map.

Usage

Theme$render_on_map(x, zindex)

Arguments

x

leaflet::leaflet() object.

zindex

numeric z-index for ordering.

Returns

leaflet::leaflet() object.

Method render_on_map()

Render on map.

Usage

Theme$render_on_map(x, zindex)

Arguments

x

leaflet::leaflet() object.

zindex

numeric z-index for ordering.

Returns

leaflet::leaflet() object.

Method update_on_map()

Render on map.

Usage

Theme$update_on_map(x, zindex)

Arguments

x

leaflet::leafletProxy() object.

zindex

numeric z-index for ordering.

Returns

leaflet::leafletProxy() object.

Method update_on_map()

Render on map.

Usage

Theme$update_on_map(x, zindex)

Arguments

x

leaflet::leafletProxy() object.

zindex

numeric z-index for ordering.

Returns

leaflet::leafletProxy() object.

Method clone()

The objects of this class are cloneable with this method.

Usage

Theme$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Method clone()

The objects of this class are cloneable with this method.

Usage

Theme$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.