Variable class

Variable class

Variable class

Details

Definition for the Variable class.

See also

Public fields

dataset

Dataset object.

index

character value.

total

numeric value.

units

character value.

legend

Legend object.

provenance

Provenance object.

Methods

Public methods

Variable$new()

Variable$print()

Variable$repr()

Variable$verify()

Variable$get_data()

Variable$export()

Variable$render()

Variable$update_render()

Variable$clone()

Method new()

Create a Variable object.

Usage

Variable$new(dataset, index, total, units, legend, provenance)

Arguments

dataset

Dataset value.

index

character or integer value.

total

numeric value.

units

character value.

legend

Legend object.

provenance

Provenance object.

Returns

A new Variable object.

Method print()

Print the object.

Usage

Variable$print(...)

Arguments

...

not used.

Method repr()

Generate a character summarizing the representation of the object.

Usage

Variable$repr(start = "[", end = "]")

Arguments

start

character symbol used to start the setting list. Defaults to "[".

end

character symbol used to start the setting list. Defaults to "]".

Returns

character value.

Method verify()

Verify that the data can be extracted from the dataset.

Usage

Variable$verify()

Returns

invisible TRUE indicating success.

Method get_data()

Get the data.

Usage

Variable$get_data()

Returns

sf::st_as_sf() or terra::rast() object.

Method export()

Export settings

Usage

Variable$export()

Returns

list object.

Method render()

Render variable on map.

Usage

Variable$render(x, id, zindex, visible)

Arguments

x

leaflet::leaflet() object.

id

character identifier for map pane.

zindex

numeric z-index for ordering.

visible

logical should the variable be visible?

Returns

leaflet::leaflet() object.

Method update_render()

Update rendering of variable on map.

Usage

Variable$update_render(x, id, zindex, visible)

Arguments

x

leaflet::leafletProxy() object.

id

character identifier for map pane.

zindex

numeric z-index for ordering.

visible

logical should the variable be visible?

Returns

leaflet::leaflet() object.

Method clone()

The objects of this class are cloneable with this method.

Usage

Variable$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.