Generate minimum set solution

Create a new Solution object by generating a prioritization using the minimum set formulation of the reserve selection problem.

min_set_solution(
  name,
  dataset,
  settings,
  theme_data = settings$get_theme_data(),
  weight_data = settings$get_weight_data(),
  include_data = settings$get_include_data(),
  boundary_data = dataset$get_boundary_data(),
  gap = 0,
  boundary_gap = 0.1,
  legend_color = "#FF0000",
  cache = cachem::cache_mem()
)

Arguments

name
character name for new solution.

dataset
Dataset object to store the solution.

settings
SolutionSettings object with settings.

theme_data
Matrix::sparseMatrix() containing the theme data. Defaults to being automatically calculated from settings.

weight_data
Matrix::sparseMatrix() containing the weight data. Defaults to being automatically calculated from settings.

include_data
Matrix::sparseMatrix() containing the include data. Defaults to being automatically calculated from settings.

boundary_data
Matrix::sparseMatrix() containing the boundary data. Defaults to being automatically extracted from dataset.

gap
numeric relative optimality gap value. Defaults to 0.

boundary_gap
numeric value used to control the level of spatial clustering in the solution. Defaults to 0.1.

legend_color
character legend color.

cache
cachem::cache_mem() object used to cache intermediate calculations. Defaults to an empty cache such that (effectively) no cache is used.

Value

A Solution object containing the solution.

Examples

# TODO.