New weight results

Create a new WeightResults object to store results for a solution.

new_weight_results(weight, held, id = uuid::UUIDgenerate())

Arguments

weight
Weight object.

held
numeric proportion of the weight covered by the solution. (e.g. 0.1 = 10%).

id
character unique identifier. Defaults to a random identifier (uuid::UUIDgenerate()).

Value

A WeightResults object.

Examples

# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
 "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
 package = "wheretowork"
)
f3 <- system.file(
 "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
 package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# create new variable
v <- new_variable_from_auto(d, index = 1)

# create a new weight
w <- new_weight(name = "NDVI", variable = v)

# create a new weight results object to store results
wr <- new_weight_results(w, 80)

# print object
print(w)
#> Weight
#>   id:       ba8d0a35-ad3f-4b85-a48b-0dd9226d3bbd
#>   name:     NDVI
#>   variable: .../sim_raster_spatial.tif#Olive_headed_Lorikeet [total: 1052 ]
#>   current:  0
#>   visible:  TRUE
#>   status:   TRUE
#>   factor:   0