Read project

Read a project from disk.

read_project(
  path,
  spatial_path = NULL,
  attribute_path = NULL,
  boundary_path = NULL,
  mode = NULL
)

Arguments

path
character file path for the configuration file.

spatial_path
character file path for spatial data. Defaults to NULL such that the file path is determined using the argument to path.

attribute_path
character file path for attribute data. Defaults to NULL such that the file path is determined using the argument to path.

boundary_path
character file path for boundary data. Defaults to NULL such that the file path is determined using the argument to path.

mode
character mode for importing the objects. Defaults to NULL such that the mode is determined based on the contents of path. If the mode is "advanced", then goal limits and mandatory include settings are disabled.

Value

A list containing the following elements:

name

A character value indicating the name.

dataset

A Dataset object.

themes

A list of Theme objects.

weights

A list of Weight objects.

includes

A list of Include objects.

mode

A character value indicating the mode.

Examples

# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_data.yaml",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f4 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# read project
x <- read_project(f1, f2, f3, f4)

# print project
print(x)
#> $name
#> [1] "Example GeoTIFF dataset"
#> 
#> $dataset
#> Dataset
#>   paths:
#>     spatial: /tmp/RtmpPRD4ri/temp_libpath1302157518e7e/wheretowork/extdata/projects/sim_raster/sim_raster_spatial.tif
#>     attribute: /tmp/RtmpPRD4ri/temp_libpath1302157518e7e/wheretowork/extdata/projects/sim_raster/sim_raster_attribute.csv.gz
#>     boundary: /tmp/RtmpPRD4ri/temp_libpath1302157518e7e/wheretowork/extdata/projects/sim_raster/sim_raster_boundary.csv.gz
#> 
#> $themes
#> $themes[[1]]
#> Theme
#>   id:        ae95afba-d685-4045-becf-69a5b998693b
#>   name:      Olive-headed Lorikeet
#>   feature: 
#>     Olive-headed Lorikeet habitat [status: TRUE, current: 0.77, goal: 0.83]
#>       variable: .../sim_raster_spatial.tif#Olive_headed_Lorikeet [total: 1052 ha]
#> 
#> $themes[[2]]
#> Theme
#>   id:        985b03da-c000-4b92-9e36-60339d589ba7
#>   name:      Purple Honeycreeper
#>   feature: 
#>     Purple Honeycreeper habitat [status: TRUE, current: 0.77, goal: 0.7]
#>       variable: .../sim_raster_spatial.tif#Purple_Honeycreeper [total: 1349 ha]
#> 
#> $themes[[3]]
#> Theme
#>   id:        2673dddd-ff16-45bd-8ba3-b434b2b278e6
#>   name:      Least Boobook
#>   feature: 
#>     Least Boobook habitat [status: TRUE, current: 0.76, goal: 0.7]
#>       variable: .../sim_raster_spatial.tif#Least_Boobook [total: 430 ha]
#> 
#> $themes[[4]]
#> Theme
#>   id:        4654fff2-5bb8-478f-8979-ab4944151b52
#>   name:      Wagtails and Pipits
#>   feature: 
#>     Eastern Yellow Wagtail (Green-headed) [status: TRUE, current: 0.78, goal: 0.69]
#>       variable: .../sim_raster_spatial.tif#Eastern_Yellow_Wagtail_Green_headed_ [total: 12825.69 ha]
#>     Alpine Pipit [status: TRUE, current: 0.78, goal: 0.55]
#>       variable: .../sim_raster_spatial.tif#Alpine_Pipit [total: 7220.08 ha]
#>     Mountain Wagtail [status: TRUE, current: 0.76, goal: 0.53]
#>       variable: .../sim_raster_spatial.tif#Mountain_Wagtail [total: 10138.58 ha]
#>     Yellow-breasted Pipit [status: TRUE, current: 0.79, goal: 0.52]
#>       variable: .../sim_raster_spatial.tif#Yellow_breasted_Pipit [total: 22013.57 ha]
#>     African Pipit (Yemen) [status: TRUE, current: 0.77, goal: 0.72]
#>       variable: .../sim_raster_spatial.tif#African_Pipit_Yemen_ [total: 21404.47 ha]
#>     Long-billed Pipit (Indian) [status: TRUE, current: 0.8, goal: 0.57]
#>       variable: .../sim_raster_spatial.tif#Long_billed_Pipit_Indian_ [total: 237.31 ha]
#>     Buffy Pipit [status: TRUE, current: 0.75, goal: 0.5]
#>       variable: .../sim_raster_spatial.tif#Buffy_Pipit [total: 2697.52 ha]
#>     Rock Pipit (Western) [status: TRUE, current: 0.78, goal: 0.54]
#>       variable: .../sim_raster_spatial.tif#Rock_Pipit_Western_ [total: 9964.19 ha]
#>     Western Yellow Wagtail (lutea) [status: TRUE, current: 0.77, goal: 0.81]
#>       variable: .../sim_raster_spatial.tif#Western_Yellow_Wagtail_lutea_ [total: 13706.62 ha]
#> 
#> $themes[[5]]
#> Theme
#>   id:        688588ea-502a-4681-abe5-5ee04cfa52b1
#>   name:      Penduline-Tits
#>   feature: 
#>     Eurasian Penduline-Tit [status: TRUE, current: 0.77, goal: 0.56]
#>       variable: .../sim_raster_spatial.tif#Eurasian_Penduline_Tit [total: 7827.21 ha]
#>     African Penduline-Tit (Slaty-backed) [status: TRUE, current: 0.8, goal: 0.56]
#>       variable: .../sim_raster_spatial.tif#African_Penduline_Tit_Slaty_backed_ [total: 16543.93 ha]
#>     Yellow Penduline-Tit [status: TRUE, current: 0.77, goal: 0.6]
#>       variable: .../sim_raster_spatial.tif#Yellow_Penduline_Tit [total: 13592.44 ha]
#>     Verdin [status: TRUE, current: 0.79, goal: 0.56]
#>       variable: .../sim_raster_spatial.tif#Verdin [total: 237.26 ha]
#>     Mouse-colored Penduline-Tit [status: TRUE, current: 0.75, goal: 0.51]
#>       variable: .../sim_raster_spatial.tif#Mouse_colored_Penduline_Tit [total: 12819.9 ha]
#> 
#> 
#> $weights
#> $weights[[1]]
#> Weight
#>   id:       6547d5d9-6a3d-424c-af80-a2900886ec45
#>   name:     Fire
#>   variable: .../sim_raster_spatial.tif#Fire [total: 147.75 ha]
#>   current:  0.77
#>   visible:  TRUE
#>   status:   TRUE
#>   factor:   0
#> 
#> $weights[[2]]
#> Weight
#>   id:       1ef04a39-0a10-4840-b2f5-cec98a7f5bfa
#>   name:     Grazing
#>   variable: .../sim_raster_spatial.tif#Grazing [total: 152.52 ha]
#>   current:  0.77
#>   visible:  TRUE
#>   status:   TRUE
#>   factor:   0
#> 
#> $weights[[3]]
#> Weight
#>   id:       5a7dce42-cefa-4271-bdb3-b636b62efc4c
#>   name:     Population count
#>   variable: .../sim_raster_spatial.tif#Population_count [total: 146.2 ha]
#>   current:  0.76
#>   visible:  TRUE
#>   status:   TRUE
#>   factor:   0
#> 
#> 
#> $includes
#> $includes[[1]]
#> Include
#>   id:       da2bceb7-9a1e-4d34-9d3d-dabe244cce88
#>   name:     OECM areas
#>   variable: .../sim_raster_spatial.tif#OECM_areas [total: 156 ]
#>   visible:  TRUE
#>   status:   TRUE
#> 
#> $includes[[2]]
#> Include
#>   id:       dd8f2879-e080-4d69-a4fe-42eeb7d764da
#>   name:     RAMSAR Sites
#>   variable: .../sim_raster_spatial.tif#RAMSAR_Sites [total: 146 ]
#>   visible:  TRUE
#>   status:   TRUE
#> 
#> 
#> $mode
#> [1] "advanced"
#>