Simulate themes

This function simulates Theme objects.

simulate_themes(dataset, n_single_themes, n_multi_themes, lambda = 5)

Arguments

dataset
Dataset object.

n_single_themes
integer number of themes with a single feature to simulate.

n_multi_themes
integer number of themes with a multiple features to simulate.

lambda
numeric lambda parameter for simulating the number of features for themes that contain multiple features. This parameter is used as a argument to rpois(n, lambda = lambda). Defaults to 5.

Value

A list of simulated Theme objects.

See also

Examples

# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# simulate data
if (requireNamespace("RandomFields")) {
 x <- simulate_themes(data = d, n_single_themes = 3, n_multi_themes = 2)
 # print results
 print(x)
}
#> Loading required namespace: RandomFields
#> .........
#> ....
#> [[1]]
#> Theme
#>   id:        519c176e-9413-42a0-9689-3068c213916d
#>   name:      Slaty-headed Parakeet
#>   feature: 
#>     Slaty-headed Parakeet habitat [status: TRUE, current: 0.49, goal: 0.54]
#>       variable: .../sim_raster_spatial.tif#Slaty_headed_Parakeet [total: 733 ha]
#> 
#> [[2]]
#> Theme
#>   id:        a9d7a5c1-610a-40dc-a73a-18ccc424bc18
#>   name:      Orange-cheeked Parrot
#>   feature: 
#>     Orange-cheeked Parrot habitat [status: TRUE, current: 0.25, goal: 0.82]
#>       variable: .../sim_raster_spatial.tif#Orange_cheeked_Parrot [total: 11662.15 ha]
#> 
#> [[3]]
#> Theme
#>   id:        a3a8def2-0b6a-4267-9913-cb53198f6b2b
#>   name:      Creamy-breasted Canastero
#>   feature: 
#>     Creamy-breasted Canastero habitat [status: TRUE, current: 0.59, goal: 0.51]
#>       variable: .../sim_raster_spatial.tif#Creamy_breasted_Canastero [total: 1534 ha]
#> 
#> [[4]]
#> Theme
#>   id:        b08c552b-7660-40d8-b028-a2f5eeed543a
#>   name:      Australasian Treecreepers
#>   feature: 
#>     Black-tailed Treecreeper [status: TRUE, current: 0.5, goal: 0.72]
#>       variable: .../sim_raster_spatial.tif#Black_tailed_Treecreeper [total: 1285 ha]
#>     White-throated Treecreeper [status: TRUE, current: 0.44, goal: 0.79]
#>       variable: .../sim_raster_spatial.tif#White_throated_Treecreeper [total: 1280 ha]
#>     Papuan Treecreeper [status: TRUE, current: 0.21, goal: 0.75]
#>       variable: .../sim_raster_spatial.tif#Papuan_Treecreeper [total: 1092 ha]
#>     Rufous Treecreeper [status: TRUE, current: 0.39, goal: 0.7]
#>       variable: .../sim_raster_spatial.tif#Rufous_Treecreeper [total: 1168 ha]
#>     Red-browed Treecreeper [status: TRUE, current: 0.6, goal: 0.74]
#>       variable: .../sim_raster_spatial.tif#Red_browed_Treecreeper [total: 1505 ha]
#>     White-throated Treecreeper (Little) [status: TRUE, current: 0.37, goal: 0.68]
#>       variable: .../sim_raster_spatial.tif#White_throated_Treecreeper_Little_ [total: 755 ha]
#>     White-browed Treecreeper [status: TRUE, current: 0.51, goal: 0.72]
#>       variable: .../sim_raster_spatial.tif#White_browed_Treecreeper [total: 696 ha]
#>     Brown Treecreeper [status: TRUE, current: 0.1, goal: 0.8]
#>       variable: .../sim_raster_spatial.tif#Brown_Treecreeper [total: 1160 ha]
#>     White-throated Treecreeper (White-throated) [status: TRUE, current: 0.57, goal: 0.59]
#>       variable: .../sim_raster_spatial.tif#White_throated_Treecreeper_White_throated_ [total: 1288 ha]
#> 
#> [[5]]
#> Theme
#>   id:        b4914969-dd95-4700-9d83-5948e6ebdb2e
#>   name:      Southern Storm-Petrels
#>   feature: 
#>     Wilson's Storm-Petrel [status: TRUE, current: 0.24, goal: 0.66]
#>       variable: .../sim_raster_spatial.tif#Wilson_s_Storm_Petrel [total: 423 ha]
#>     White-bellied Storm-Petrel (White-bellied) [status: TRUE, current: 0.26, goal: 0.87]
#>       variable: .../sim_raster_spatial.tif#White_bellied_Storm_Petrel_White_bellied_ [total: 1700 ha]
#>     New Zealand Storm-Petrel [status: TRUE, current: 0.13, goal: 0.77]
#>       variable: .../sim_raster_spatial.tif#New_Zealand_Storm_Petrel [total: 762 ha]
#>     White-faced Storm-Petrel [status: TRUE, current: 0.28, goal: 0.6]
#>       variable: .../sim_raster_spatial.tif#White_faced_Storm_Petrel [total: 613 ha]
#>