wheretowork: Systematic conservation planning application.

The application is a decision support tool to help prioritize conservation efforts for the Nature Conservancy of Canada. It provides an interactive interface for conducting systematic conservation planning exercises, and uses mathematical optimization algorithms to generate solutions.

See also

Author

Maintainer: Jeffrey O Hanson jeffrey.hanson@uqconnect.edu.au (ORCID)

Authors:

Richard Schuster richard.schuster@glel.carleton.ca (ORCID)

Joseph Bennett

Jaimie Vincent

Other contributors:

Dan Wismer dan.wismer@natureconservancy.ca [contributor]

Marc Edwards marc.edwards@natureconservancy.ca [contributor]

Examples

# \donttest{
if (interactive()) {
run_app()
}
# }