Read project

Read a project from disk.

read_project(
  path,
  spatial_path = NULL,
  attribute_path = NULL,
  boundary_path = NULL,
  mode = "project"
)

Arguments

path
character file path for the configuration file.

spatial_path
character file path for spatial data. Defaults to NULL such that the file path is determined using the argument to path.

attribute_path
character file path for attribute data. Defaults to NULL such that the file path is determined using the argument to path.

boundary_path
character file path for boundary data. Defaults to NULL such that the file path is determined using the argument to path.

mode
character mode for importing the objects. Defaults to "project" such that the mode is determined based on the contents of path. If the mode is "advanced", then goal limits and mandatory include settings are disabled.

Value

A list containing the following elements:

name

A character value indicating the name.

dataset

A Dataset object.

themes

A list of Theme objects.

weights

A list of Weight objects.

includes

A list of Include objects.

mode

A character value indicating the mode.

Details

Note that the status field for themes and weights will automatically be set to FALSE if the goal/weight values are zero. This is to ensure that the application starts with a sensible combination of settings.

Examples

# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_data.yaml",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f4 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# read project
x <- read_project(f1, f2, f3, f4)

# print project
print(x)
#> $name
#> [1] "Example GeoTIFF dataset"
#> 
#> $author_name
#> [1] "Richard Schuster"
#> 
#> $author_email
#> [1] "richard.schuster@natureconservancy.ca"
#> 
#> $dataset
#> Dataset
#>   paths:
#>     spatial: /tmp/RtmpGyKZqg/temp_libpath9ff641df8a01/wheretowork/extdata/projects/sim_raster/sim_raster_spatial.tif
#>     attribute: /tmp/RtmpGyKZqg/temp_libpath9ff641df8a01/wheretowork/extdata/projects/sim_raster/sim_raster_attribute.csv.gz
#>     boundary: /tmp/RtmpGyKZqg/temp_libpath9ff641df8a01/wheretowork/extdata/projects/sim_raster/sim_raster_boundary.csv.gz
#> 
#> $themes
#> $themes[[1]]
#> Theme
#>   id:        a9a76fb3-e67f-4b76-bdd2-e144fdd7b9c0
#>   name:      Brown-necked Parrot (Brown-necked)
#>   feature: 
#>     Brown-necked Parrot (Brown-necked) habitat [status: TRUE, current: 0.78, goal: 0.83]
#>       variable: .../sim_raster_spatial.tif#Brown_necked_Parrot_Brown_necked_ [total: 1052 ha]
#> 
#> $themes[[2]]
#> Theme
#>   id:        553292d8-38c3-480b-acf3-b5848f8e569b
#>   name:      Gough Island Finch
#>   feature: 
#>     Gough Island Finch habitat [status: TRUE, current: 0.78, goal: 0.7]
#>       variable: .../sim_raster_spatial.tif#Gough_Island_Finch [total: 1349 ha]
#> 
#> $themes[[3]]
#> Theme
#>   id:        be1c447c-9014-4142-b970-299fff9940eb
#>   name:      Manus Boobook
#>   feature: 
#>     Manus Boobook habitat [status: TRUE, current: 0.78, goal: 0.7]
#>       variable: .../sim_raster_spatial.tif#Manus_Boobook [total: 430 ha]
#> 
#> $themes[[4]]
#> Theme
#>   id:        e8c02818-eae6-4b8a-8cb7-38cc7f2ef0de
#>   name:      Waxbills and Allies
#>   feature: 
#>     Red-faced Crimsonwing [status: TRUE, current: 0.75, goal: 0.69]
#>       variable: .../sim_raster_spatial.tif#Red_faced_Crimsonwing [total: 12825.69 ha]
#>     African Firefinch [status: TRUE, current: 0.76, goal: 0.55]
#>       variable: .../sim_raster_spatial.tif#African_Firefinch [total: 7220.08 ha]
#>     Black-and-white Mannikin (Red-backed) [status: TRUE, current: 0.76, goal: 0.53]
#>       variable: .../sim_raster_spatial.tif#Black_and_white_Mannikin_Red_backed_ [total: 10138.58 ha]
#>     Black-throated Finch (White-rumped) [status: TRUE, current: 0.79, goal: 0.72]
#>       variable: .../sim_raster_spatial.tif#Black_throated_Finch_White_rumped_ [total: 22013.57 ha]
#>     Crimson Seedcracker [status: TRUE, current: 0.79, goal: 0.61]
#>       variable: .../sim_raster_spatial.tif#Crimson_Seedcracker [total: 21404.47 ha]
#>     Royal Parrotfinch [status: TRUE, current: 0.77, goal: 0.54]
#>       variable: .../sim_raster_spatial.tif#Royal_Parrotfinch [total: 237.31 ha]
#>     Red-billed Firefinch [status: TRUE, current: 0.8, goal: 0.81]
#>       variable: .../sim_raster_spatial.tif#Red_billed_Firefinch [total: 2697.52 ha]
#>     Quailfinch (Black-faced) [status: TRUE, current: 0.77, goal: 0.76]
#>       variable: .../sim_raster_spatial.tif#Quailfinch_Black_faced_ [total: 9964.19 ha]
#>     White-collared Oliveback [status: TRUE, current: 0.8, goal: 0.86]
#>       variable: .../sim_raster_spatial.tif#White_collared_Oliveback [total: 13706.62 ha]
#> 
#> $themes[[5]]
#> Theme
#>   id:        626fa34b-62b2-4536-ac52-bc9e3a06be1a
#>   name:      Penduline-Tits
#>   feature: 
#>     Eurasian Penduline-Tit [status: TRUE, current: 0.77, goal: 0.8]
#>       variable: .../sim_raster_spatial.tif#Eurasian_Penduline_Tit [total: 5229.66 ha]
#>     Verdin [status: TRUE, current: 0.81, goal: 0.64]
#>       variable: .../sim_raster_spatial.tif#Verdin [total: 16793.02 ha]
#>     Mouse-colored Penduline-Tit [status: TRUE, current: 0.75, goal: 0.66]
#>       variable: .../sim_raster_spatial.tif#Mouse_colored_Penduline_Tit [total: 15611.47 ha]
#>     Forest Penduline-Tit [status: TRUE, current: 0.74, goal: 0.68]
#>       variable: .../sim_raster_spatial.tif#Forest_Penduline_Tit [total: 14553.23 ha]
#>     Chinese Penduline-Tit [status: TRUE, current: 0.78, goal: 0.8]
#>       variable: .../sim_raster_spatial.tif#Chinese_Penduline_Tit [total: 5295 ha]
#> 
#> 
#> $weights
#> $weights[[1]]
#> Weight
#>   id:       2926ddae-e608-490c-9194-9ee21a32104f
#>   name:     Landscape naturalness
#>   variable: .../sim_raster_spatial.tif#Landscape_naturalness [total: 146.73 ha]
#>   current:  0.78
#>   visible:  TRUE
#>   hidden:  FALSE
#>   status:   FALSE
#>   factor:   0
#> 
#> $weights[[2]]
#> Weight
#>   id:       e181b04f-16f7-4ae2-b7dc-1025dbec715d
#>   name:     Archaeological site
#>   variable: .../sim_raster_spatial.tif#Archaeological_site [total: 153.11 ha]
#>   current:  0.78
#>   visible:  TRUE
#>   hidden:  FALSE
#>   status:   FALSE
#>   factor:   0
#> 
#> $weights[[3]]
#> Weight
#>   id:       5033d464-a300-45af-a919-fb18826c802c
#>   name:     Engagement potential
#>   variable: .../sim_raster_spatial.tif#Engagement_potential [total: 145.9 ha]
#>   current:  0.78
#>   visible:  TRUE
#>   hidden:  FALSE
#>   status:   FALSE
#>   factor:   0
#> 
#> 
#> $includes
#> $includes[[1]]
#> Include
#>   id:       ffb4db02-201c-4d0b-8bc1-1a695d496a7d
#>   name:     National protected areas
#>   variable: .../sim_raster_spatial.tif#National_protected_areas [total: 159 ]
#>   visible:  TRUE
#>   hidden:  FALSE
#>   status:   TRUE
#> 
#> $includes[[2]]
#> Include
#>   id:       3cb131b6-1251-46be-b408-17600ede0c25
#>   name:     Key Biodiversity Areas
#>   variable: .../sim_raster_spatial.tif#Key_Biodiversity_Areas [total: 147 ]
#>   visible:  TRUE
#>   hidden:  FALSE
#>   status:   TRUE
#> 
#> 
#> $mode
#> [1] "advanced"
#>