Simulate themes

This function simulates Theme objects.

simulate_themes(dataset, n_single_themes, n_multi_themes, lambda = 5)

Arguments

dataset
Dataset object.

n_single_themes
integer number of themes with a single feature to simulate.

n_multi_themes
integer number of themes with a multiple features to simulate.

lambda
numeric lambda parameter for simulating the number of features for themes that contain multiple features. This parameter is used as a argument to rpois(n, lambda = lambda). Defaults to 5.

Value

A list of simulated Theme objects.

See also

Examples

# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# simulate data
x <- simulate_themes(data = d, n_single_themes = 3, n_multi_themes = 2)
#> .......
#> ..

# print results
print(x)
#> [[1]]
#> Theme
#>   id:        8283bca6-6d14-4b8c-874b-b4780210ac64
#>   name:      Sooty Grassquit
#>   feature: 
#>     Sooty Grassquit habitat [status: TRUE, current: 0.36, goal: 0.78]
#>       variable: .../sim_raster_spatial.tif#Sooty_Grassquit [total: 420 ha]
#> 
#> [[2]]
#> Theme
#>   id:        214236a1-4473-460b-9136-03a18a2aa0ef
#>   name:      Redwing
#>   feature: 
#>     Redwing habitat [status: TRUE, current: 0.26, goal: 0.52]
#>       variable: .../sim_raster_spatial.tif#Redwing [total: 611 ha]
#> 
#> [[3]]
#> Theme
#>   id:        9aa773bb-c270-48bb-a1e0-5809888fc40d
#>   name:      Maroon-tailed Parakeet
#>   feature: 
#>     Maroon-tailed Parakeet habitat [status: TRUE, current: 0.43, goal: 0.75]
#>       variable: .../sim_raster_spatial.tif#Maroon_tailed_Parakeet [total: 22066.78 ha]
#> 
#> [[4]]
#> Theme
#>   id:        61511bb4-7565-479c-b8a5-2a16fc69d4c8
#>   name:      Thick-knees
#>   feature: 
#>     Double-striped Thick-knee [status: TRUE, current: 0.19, goal: 0.51]
#>       variable: .../sim_raster_spatial.tif#Double_striped_Thick_knee [total: 1104 ha]
#>     Bush Thick-knee [status: TRUE, current: 0.27, goal: 0.83]
#>       variable: .../sim_raster_spatial.tif#Bush_Thick_knee [total: 434 ha]
#>     Great Thick-knee [status: TRUE, current: 0.42, goal: 0.76]
#>       variable: .../sim_raster_spatial.tif#Great_Thick_knee [total: 1343 ha]
#>     Indian Thick-knee [status: TRUE, current: 0.36, goal: 0.83]
#>       variable: .../sim_raster_spatial.tif#Indian_Thick_knee [total: 774 ha]
#>     Peruvian Thick-knee [status: TRUE, current: 0.58, goal: 0.6]
#>       variable: .../sim_raster_spatial.tif#Peruvian_Thick_knee [total: 1448 ha]
#>     Water Thick-knee [status: TRUE, current: 0.54, goal: 0.85]
#>       variable: .../sim_raster_spatial.tif#Water_Thick_knee [total: 1412 ha]
#>     Beach Thick-knee [status: TRUE, current: 0.47, goal: 0.76]
#>       variable: .../sim_raster_spatial.tif#Beach_Thick_knee [total: 733 ha]
#> 
#> [[5]]
#> Theme
#>   id:        0819aecf-26b5-4fb5-9cb4-98cec47fa346
#>   name:      Waxwings
#>   feature: 
#>     Japanese Waxwing [status: TRUE, current: 0.33, goal: 0.74]
#>       variable: .../sim_raster_spatial.tif#Japanese_Waxwing [total: 20960.1 ha]
#>     Bohemian Waxwing [status: TRUE, current: 0.27, goal: 0.62]
#>       variable: .../sim_raster_spatial.tif#Bohemian_Waxwing [total: 13696.99 ha]
#>