Simulate themes

This function simulates Theme objects.

simulate_themes(dataset, n_single_themes, n_multi_themes, lambda = 5)

Arguments

dataset
Dataset object.

n_single_themes
integer number of themes with a single feature to simulate.

n_multi_themes
integer number of themes with a multiple features to simulate.

lambda
numeric lambda parameter for simulating the number of features for themes that contain multiple features. This parameter is used as a argument to rpois(n, lambda = lambda). Defaults to 5.

Value

A list of simulated Theme objects.

See also

Examples

# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# simulate data
x <- simulate_themes(data = d, n_single_themes = 3, n_multi_themes = 2)
#> ...
#> .........

# print results
print(x)
#> [[1]]
#> Theme
#>   id:        d946c5b1-7559-4e8c-8773-61e1cd213102
#>   name:      Mountain Cacique (Bolivian)
#>   feature: 
#>     Mountain Cacique (Bolivian) habitat [status: TRUE, current: 0.21, goal: 0.51]
#>       variable: .../sim_raster_spatial.tif#Mountain_Cacique_Bolivian_ [total: 17016.42 ha]
#> 
#> [[2]]
#> Theme
#>   id:        a002f698-de9c-4c2e-8095-3da31bed68f6
#>   name:      Spotted Imperial-Pigeon
#>   feature: 
#>     Spotted Imperial-Pigeon habitat [status: TRUE, current: 0.42, goal: 0.78]
#>       variable: .../sim_raster_spatial.tif#Spotted_Imperial_Pigeon [total: 1585 ha]
#> 
#> [[3]]
#> Theme
#>   id:        d7b974ca-02c0-4ae8-9bd1-a9e579c5ee18
#>   name:      Spotted Ground-Thrush (Upemba)
#>   feature: 
#>     Spotted Ground-Thrush (Upemba) habitat [status: TRUE, current: 0.37, goal: 0.7]
#>       variable: .../sim_raster_spatial.tif#Spotted_Ground_Thrush_Upemba_ [total: 11650.44 ha]
#> 
#> [[4]]
#> Theme
#>   id:        3e55c71e-ffae-4088-bb5e-5ab07fa25cbc
#>   name:      Kiwis
#>   feature: 
#>     Southern Brown Kiwi (South I.) [status: TRUE, current: 0.15, goal: 0.5]
#>       variable: .../sim_raster_spatial.tif#Southern_Brown_Kiwi_South_I_ [total: 2562.48 ha]
#>     Little Spotted Kiwi [status: TRUE, current: 0.41, goal: 0.79]
#>       variable: .../sim_raster_spatial.tif#Little_Spotted_Kiwi [total: 9734.59 ha]
#>     North Island Brown Kiwi [status: TRUE, current: 0.47, goal: 0.85]
#>       variable: .../sim_raster_spatial.tif#North_Island_Brown_Kiwi [total: 242.12 ha]
#> 
#> [[5]]
#> Theme
#>   id:        c46f6bf6-ba38-452b-b8bf-f4e3030013b5
#>   name:      Jacamars
#>   feature: 
#>     White-throated Jacamar [status: TRUE, current: 0.34, goal: 0.8]
#>       variable: .../sim_raster_spatial.tif#White_throated_Jacamar [total: 2660.92 ha]
#>     Dusky-backed Jacamar [status: TRUE, current: 0.26, goal: 0.85]
#>       variable: .../sim_raster_spatial.tif#Dusky_backed_Jacamar [total: 17298.69 ha]
#>     Purus Jacamar [status: TRUE, current: 0.5, goal: 0.75]
#>       variable: .../sim_raster_spatial.tif#Purus_Jacamar [total: 9017.74 ha]
#>     White-chinned Jacamar [status: TRUE, current: 0.3, goal: 0.54]
#>       variable: .../sim_raster_spatial.tif#White_chinned_Jacamar [total: 12824.3 ha]
#>     Brown Jacamar [status: TRUE, current: 0.45, goal: 0.89]
#>       variable: .../sim_raster_spatial.tif#Brown_Jacamar [total: 17489.43 ha]
#>     Yellow-billed Jacamar (Cerise-crowned) [status: TRUE, current: 0.13, goal: 0.81]
#>       variable: .../sim_raster_spatial.tif#Yellow_billed_Jacamar_Cerise_crowned_ [total: 16295.13 ha]
#>     Three-toed Jacamar [status: TRUE, current: 0.44, goal: 0.81]
#>       variable: .../sim_raster_spatial.tif#Three_toed_Jacamar [total: 17214.64 ha]
#>     Rufous-tailed Jacamar (Spot-tailed) [status: TRUE, current: 0.52, goal: 0.74]
#>       variable: .../sim_raster_spatial.tif#Rufous_tailed_Jacamar_Spot_tailed_ [total: 14469.06 ha]
#>     Paradise Jacamar [status: TRUE, current: 0.48, goal: 0.71]
#>       variable: .../sim_raster_spatial.tif#Paradise_Jacamar [total: 7347.9 ha]
#>