Simulate themes

This function simulates Theme objects.

simulate_themes(
  dataset,
  n_single_themes,
  n_multi_themes,
  lambda = 5,
  continuous = NA
)

Arguments

dataset
Dataset object.

n_single_themes
integer number of themes with a single feature to simulate.

n_multi_themes
integer number of themes with a multiple features to simulate.

lambda
numeric lambda parameter for simulating the number of features for themes that contain multiple features. This parameter is used as a argument to rpois(n, lambda = lambda). Defaults to 5.

continuous
logical should the data be continuous? Defaults to NA. If NA, simulated themes are a mix of continuous and categorical defined by a conditional: stats::runif(1) > 0.5

Value

A list of simulated Theme objects.

See also

Examples

# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# simulate data
if (requireNamespace("RandomFields")) {
 x <- simulate_themes(data = d, n_single_themes = 3, n_multi_themes = 2)
 # print results
 print(x)
}
#> ......
#> ...
#> [[1]]
#> Theme
#>   id:        0b19001a-b974-4e5b-a125-cc2120159e30
#>   name:      Dusky-billed Parrotlet
#>   feature: 
#>     Dusky-billed Parrotlet habitat [status: TRUE, current: 0.15, goal: 0.58]
#>       variable: .../sim_raster_spatial.tif#Dusky_billed_Parrotlet [total: 799 ha]
#> 
#> [[2]]
#> Theme
#>   id:        9e9e3cab-af0d-419e-8f11-e4a0fad9c054
#>   name:      Red-bellied Fruit-Dove
#>   feature: 
#>     Red-bellied Fruit-Dove habitat [status: TRUE, current: 0.14, goal: 0.62]
#>       variable: .../sim_raster_spatial.tif#Red_bellied_Fruit_Dove [total: 9851.68 ha]
#> 
#> [[3]]
#> Theme
#>   id:        09394ed2-80e9-4b1f-a3c1-9a60d72ff6bc
#>   name:      Kentish Plover (Indian)
#>   feature: 
#>     Kentish Plover (Indian) habitat [status: TRUE, current: 0.53, goal: 0.53]
#>       variable: .../sim_raster_spatial.tif#Kentish_Plover_Indian_ [total: 1122 ha]
#> 
#> [[4]]
#> Theme
#>   id:        f8810177-285e-4c4f-a294-a6d5b4ef70eb
#>   name:      Woodswallows, Bellmagpies, and Allies
#>   feature: 
#>     Silver-backed Butcherbird [status: TRUE, current: 0.6, goal: 0.8]
#>       variable: .../sim_raster_spatial.tif#Ivory_backed_Woodswallow [total: 455 ha]
#>     Black Butcherbird [status: TRUE, current: 0.33, goal: 0.65]
#>       variable: .../sim_raster_spatial.tif#Ivory_backed_Woodswallow [total: 455 ha]
#>     Dusky Woodswallow [status: TRUE, current: 0.33, goal: 0.76]
#>       variable: .../sim_raster_spatial.tif#Ivory_backed_Woodswallow [total: 455 ha]
#>     Bismarck Woodswallow [status: TRUE, current: 0.19, goal: 0.73]
#>       variable: .../sim_raster_spatial.tif#Ivory_backed_Woodswallow [total: 455 ha]
#>     Gray Currawong (Clinking) [status: TRUE, current: 0.55, goal: 0.89]
#>       variable: .../sim_raster_spatial.tif#Ivory_backed_Woodswallow [total: 455 ha]
#>     Ivory-backed Woodswallow [status: TRUE, current: 0.44, goal: 0.53]
#>       variable: .../sim_raster_spatial.tif#Ivory_backed_Woodswallow [total: 455 ha]
#> 
#> [[5]]
#> Theme
#>   id:        7d26f482-f0f9-4920-905e-9b9c6c978acf
#>   name:      Indigobirds
#>   feature: 
#>     Wilson's Indigobird [status: TRUE, current: 0.24, goal: 0.84]
#>       variable: .../sim_raster_spatial.tif#Baka_Indigobird [total: 21535.41 ha]
#>     Togo Paradise-Whydah [status: TRUE, current: 0.11, goal: 0.74]
#>       variable: .../sim_raster_spatial.tif#Baka_Indigobird [total: 21535.41 ha]
#>     Baka Indigobird [status: TRUE, current: 0.12, goal: 0.67]
#>       variable: .../sim_raster_spatial.tif#Baka_Indigobird [total: 21535.41 ha]
#>