Simulate themes

This function simulates SingleTheme and MultiTheme objects.

simulate_themes(dataset, n_single_themes, n_multi_themes, lambda = 5)

Arguments

dataset
Dataset object.

n_single_themes
integer number of themes with a single feature to simulate.

n_multi_themes
integer number of themes with a multiple features to simulate.

lambda
numeric lambda parameter for simulating the number of features for each MultiTheme object. This parameter is used as a argument to rpois(n, lambda = lambda). Defaults to 5.

Value

A list of simulated Theme objects.

See also

Examples

# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
 "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
 package = "wheretowork"
)
f3 <- system.file(
 "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
 package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# simulate data
x <- simulate_themes(data = d, n_single_themes = 3, n_multi_themes = 2)
#> ....
#> ......

# print results
print(x)
#> [[1]]
#> Theme
#>   id:        c4b3ffab-f8e1-4dd1-b1ef-1c4b121d94eb
#>   name:      Asian Golden Weaver
#>   feature: 
#>     Asian Golden Weaver habitat [status: TRUE, current: 0.12, goal: 0.67]
#>       variable: .../sim_raster_spatial.tif#Asian_Golden_Weaver [total: 231.55 ha]
#> 
#> [[2]]
#> Theme
#>   id:        ad7b6c87-d767-4538-9b28-6e988ee773a2
#>   name:      Mountain Chiffchaff
#>   feature: 
#>     Mountain Chiffchaff habitat [status: TRUE, current: 0.21, goal: 0.73]
#>       variable: .../sim_raster_spatial.tif#Mountain_Chiffchaff [total: 803 ha]
#> 
#> [[3]]
#> Theme
#>   id:        91c4ac09-407e-459b-a413-c8c8a2ffbd53
#>   name:      Cinereous Warbling-Finch
#>   feature: 
#>     Cinereous Warbling-Finch habitat [status: TRUE, current: 0.54, goal: 0.71]
#>       variable: .../sim_raster_spatial.tif#Cinereous_Warbling_Finch [total: 1133 ha]
#> 
#> [[4]]
#> Theme
#>   id:        8e95f6ce-d2ec-463a-999c-b9d8991f557b
#>   name:      Gulls, Terns, and Skimmers
#>   feature: 
#>     Laughing Gull [status: TRUE, current: 0.14, goal: 0.51]
#>       variable: .../sim_raster_spatial.tif#Laughing_Gull [total: 10167.06 ha]
#>     Franklin's Gull [status: TRUE, current: 0.3, goal: 0.64]
#>       variable: .../sim_raster_spatial.tif#Franklin_s_Gull [total: 9379.63 ha]
#>     Iceland Gull (Thayer's) [status: TRUE, current: 0.25, goal: 0.72]
#>       variable: .../sim_raster_spatial.tif#Iceland_Gull_Thayer_s_ [total: 18754.55 ha]
#>     Lesser Black-backed Gull [status: TRUE, current: 0.53, goal: 0.89]
#>       variable: .../sim_raster_spatial.tif#Lesser_Black_backed_Gull [total: 19482.9 ha]
#> 
#> [[5]]
#> Theme
#>   id:        8b626b19-6d06-45fc-bd07-41268440badf
#>   name:      Bush Warblers and Allies
#>   feature: 
#>     Brownish-flanked Bush Warbler (Taiwan) [status: TRUE, current: 0.55, goal: 0.63]
#>       variable: .../sim_raster_spatial.tif#Brownish_flanked_Bush_Warbler_Taiwan_ [total: 13579.89 ha]
#>     Fiji Bush Warbler [status: TRUE, current: 0.28, goal: 0.76]
#>       variable: .../sim_raster_spatial.tif#Fiji_Bush_Warbler [total: 10965.74 ha]
#>     Slaty-bellied Tesia [status: TRUE, current: 0.59, goal: 0.78]
#>       variable: .../sim_raster_spatial.tif#Slaty_bellied_Tesia [total: 20785.6 ha]
#>     Hume's Bush Warbler [status: TRUE, current: 0.48, goal: 0.9]
#>       variable: .../sim_raster_spatial.tif#Hume_s_Bush_Warbler [total: 7212.96 ha]
#>     Brownish-flanked Bush Warbler (Brownish-flanked) [status: TRUE, current: 0.57, goal: 0.67]
#>       variable: .../sim_raster_spatial.tif#Brownish_flanked_Bush_Warbler_Brownish_flanked_ [total: 17523.29 ha]
#>     Broad-billed Warbler [status: TRUE, current: 0.21, goal: 0.85]
#>       variable: .../sim_raster_spatial.tif#Broad_billed_Warbler [total: 7280.45 ha]
#>