Simulate themes

This function simulates Theme objects.

simulate_themes(
  dataset,
  n_single_themes,
  n_multi_themes,
  lambda = 5,
  continuous = NA
)

Arguments

dataset
Dataset object.

n_single_themes
integer number of themes with a single feature to simulate.

n_multi_themes
integer number of themes with a multiple features to simulate.

lambda
numeric lambda parameter for simulating the number of features for themes that contain multiple features. This parameter is used as a argument to rpois(n, lambda = lambda). Defaults to 5.

continuous
logical should the data be continuous? Defaults to NA. If NA, simulated themes are a mix of continuous and categorical defined by a conditional: stats::runif(1) > 0.5

Value

A list of simulated Theme objects.

See also

Examples

# find data file paths
f1 <- system.file(
  "extdata", "projects", "sim_raster", "sim_raster_spatial.tif",
  package = "wheretowork"
)
f2 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_attribute.csv.gz",
  package = "wheretowork"
)
f3 <- system.file(
  "extdata",  "projects", "sim_raster", "sim_raster_boundary.csv.gz",
  package = "wheretowork"
)

# create new dataset
d <- new_dataset(f1, f2, f3)

# simulate data
if (requireNamespace("RandomFields")) {
 x <- simulate_themes(data = d, n_single_themes = 3, n_multi_themes = 2)
 # print results
 print(x)
}
#> .....
#> .......
#> [[1]]
#> Theme
#>   id:        d6db9fcc-4c5b-4a11-998d-a2ff3ea5bbfc
#>   name:      Eastern Yellow-billed Hornbill
#>   feature: 
#>     Eastern Yellow-billed Hornbill habitat [status: TRUE, current: 0.47, goal: 0.78]
#>       variable: .../sim_raster_spatial.tif#Eastern_Yellow_billed_Hornbill [total: 2543.68 ha]
#> 
#> [[2]]
#> Theme
#>   id:        cbd9d8e9-4896-44f1-97e9-c32d148c2ddb
#>   name:      Collared Inca (Gould's)
#>   feature: 
#>     Collared Inca (Gould's) habitat [status: TRUE, current: 0.44, goal: 0.64]
#>       variable: .../sim_raster_spatial.tif#Collared_Inca_Gould_s_ [total: 9971.96 ha]
#> 
#> [[3]]
#> Theme
#>   id:        c53e6c4b-0462-437b-910e-eb94d5d77546
#>   name:      Chestnut-breasted Partridge
#>   feature: 
#>     Chestnut-breasted Partridge habitat [status: TRUE, current: 0.49, goal: 0.87]
#>       variable: .../sim_raster_spatial.tif#Chestnut_breasted_Partridge [total: 1217 ha]
#> 
#> [[4]]
#> Theme
#>   id:        7121ebfb-e0fe-454a-9828-bd2ec94f6239
#>   name:      Cuckoos
#>   feature: 
#>     Dwarf Koel [status: TRUE, current: 0.46, goal: 0.7]
#>       variable: .../sim_raster_spatial.tif#Crested_Coua_Crested_ [total: 17752.16 ha]
#>     Little Bronze-Cuckoo (Banda) [status: TRUE, current: 0.6, goal: 0.69]
#>       variable: .../sim_raster_spatial.tif#Crested_Coua_Crested_ [total: 17752.16 ha]
#>     Gabon Coucal [status: TRUE, current: 0.38, goal: 0.84]
#>       variable: .../sim_raster_spatial.tif#Crested_Coua_Crested_ [total: 17752.16 ha]
#>     Running Coua [status: TRUE, current: 0.53, goal: 0.9]
#>       variable: .../sim_raster_spatial.tif#Crested_Coua_Crested_ [total: 17752.16 ha]
#>     Crested Coua (Crested) [status: TRUE, current: 0.4, goal: 0.6]
#>       variable: .../sim_raster_spatial.tif#Crested_Coua_Crested_ [total: 17752.16 ha]
#> 
#> [[5]]
#> Theme
#>   id:        79eaa247-7dc6-4579-a78d-a648ba8db336
#>   name:      Cormorants and Shags
#>   feature: 
#>     Red-faced Cormorant [status: TRUE, current: 0.43, goal: 0.74]
#>       variable: .../sim_raster_spatial.tif#European_Shag_Atlantic_ [total: 726 ha]
#>     Flightless Cormorant [status: TRUE, current: 0.41, goal: 0.8]
#>       variable: .../sim_raster_spatial.tif#European_Shag_Atlantic_ [total: 726 ha]
#>     Pelagic Cormorant [status: TRUE, current: 0.59, goal: 0.67]
#>       variable: .../sim_raster_spatial.tif#European_Shag_Atlantic_ [total: 726 ha]
#>     Pied Cormorant [status: TRUE, current: 0.43, goal: 0.68]
#>       variable: .../sim_raster_spatial.tif#European_Shag_Atlantic_ [total: 726 ha]
#>     Campbell Islands Shag [status: TRUE, current: 0.29, goal: 0.8]
#>       variable: .../sim_raster_spatial.tif#European_Shag_Atlantic_ [total: 726 ha]
#>     European Shag (Moroccan) [status: TRUE, current: 0.32, goal: 0.66]
#>       variable: .../sim_raster_spatial.tif#European_Shag_Atlantic_ [total: 726 ha]
#>     European Shag (Atlantic) [status: TRUE, current: 0.29, goal: 0.58]
#>       variable: .../sim_raster_spatial.tif#European_Shag_Atlantic_ [total: 726 ha]
#>